LPS-49438 remove abstractions
Wait until we need to translate Journal content types to reintroduce
them.
